+++ Faker_ip_v4_segment.v
@@ -1,5 +1,5 @@
 module faker
 
 fn (mut this Faker) ip_v4_segment() u8 {
-    return this.randomizer.u8() % 256
+    return u8(this.randomizer.u8() % 256)
 }
